### 2025-08-15 #8797 @cansavvy
> @cansavvy - thanks for your submission to JOSS. We're currently managing a large backlog of submissions and the editors most appropriate for your area are all already rather busy.
>
> For now, we will need to waitlist this paper and process it as the queue reduces. Thanks for your patience!

No worries. Thanks for the update!
