### 2025-12-05 229150e
Make downloads happen gracefully
