+++ R/03-annotate.R
@@ -139,6 +139,10 @@ gimap_annotate <- function(.data = NULL,
     annotation_df <- readr::read_table(annotation_file)
   } else {
     annotation_df <- get_example_data("annotation")
+    if (is.null(annotation_df)) {
+      message("Could not retrieve annotation data. Returning dataset without annotation.")
+      return(gimap_dataset)
+    }
   }
 
   message("Annotating Data")
@@ -159,16 +163,37 @@ gimap_annotate <- function(.data = NULL,
     # Essential gene labeling is from
     # inst/extdata/Achilles_common_essentials.csv
     control_genes <- ctrl_genes()
+    if (is.null(control_genes)) {
+      message("Could not retrieve control genes. Returning dataset without annotation.")
+      return(gimap_dataset)
+    }
   }
 
   ############################ Get TPM data ####################################
   if (cell_line_annotate) {
     # This is used to flag things
     ## get TPM and CN information (w/ option for user to upload their own info)
-    depmap_metadata <- readr::read_csv(
-      "https://figshare.com/ndownloader/files/35020903",
-      show_col_types = FALSE
+    depmap_metadata <- tryCatch(
+      {
+        readr::read_csv(
+          "https://figshare.com/ndownloader/files/35020903",
+          show_col_types = FALSE
+        )
+      },
+      error = function(e) {
+        message(
+          "Could not download DepMap metadata. ",
+          "Please check your internet connection and try again later.\n",
+          "Error: ", e$message
+        )
+        return(NULL)
+      }
     )
+
+    if (is.null(depmap_metadata)) {
+      message("Returning dataset without cell line annotation due to network error.")
+      return(gimap_dataset)
+    }
 
     my_depmap_id <- depmap_metadata %>%
       dplyr::filter(stripped_cell_line_name == toupper(cell_line)) %>%
@@ -184,7 +209,10 @@ gimap_annotate <- function(.data = NULL,
     }
     tpm_file <- getOption("tpm_file")
     if (is.null(tpm_file)) tpm_file <- tpm_setup(data_dir = annot_dir)
-    if (!file.exists(tpm_file)) stop("Could not download TPM file")
+    if (is.null(tpm_file) || !file.exists(tpm_file)) {
+      message("Could not download TPM file. Returning dataset without cell line annotation.")
+      return(gimap_dataset)
+    }
 
     op <- options("VROOM_CONNECTION_SIZE" = 500072)
     on.exit(options(op))
@@ -200,7 +228,10 @@ gimap_annotate <- function(.data = NULL,
 
     if (is.null(cn_file)) cn_file <- cn_setup(data_dir = annot_dir)
 
-    if (!file.exists(cn_file)) stop("Could not download CN file")
+    if (is.null(cn_file) || !file.exists(cn_file)) {
+      message("Could not download CN file. Returning dataset without cell line annotation.")
+      return(gimap_dataset)
+    }
     # Read in the CN data
     depmap_cn <- readr::read_csv(cn_file,
       show_col_types = FALSE,
@@ -343,10 +374,14 @@ tpm_setup <- function(overwrite = TRUE,
   options("tpm_file" = tpm_file)
   if (!file.exists(tpm_file) | overwrite) {
     if (!file.exists(file.path(data_dir, "CCLE_expression.csv.zip"))) {
-      get_figshare(
+      download_result <- get_figshare(
         file_name = "CCLE_expression.csv",
         output_dir = data_dir
       )
+      if (is.null(download_result)) {
+        message("Could not download TPM data from Figshare.")
+        return(NULL)
[…]
