### 2025-12-05 f00caff
one sample t test if use_ntc = FALSE
