### 2025-12-05 f3bd355
Fix test handling and docs
