+++ README.md
@@ -1,4 +1,7 @@
 # We're adding a new line to test a pull request
+
+# And now adding a second lin
+
 # gimap - Genetic Interaction MAPping for dual target CRISPR screens
 
 <!-- START doctoc generated TOC please keep comment here to allow auto update -->
