### 2026-01-06 #8797 @rkurchin
Hi @cansavvy, BCM track AEiC here!

As we work to find an editor for your submission, I wanted to give you a heads-up that we have recently updated our scope criteria, so take note of the announcement [here](https://blog.joss.theoj.org/2026/01/preparing-joss-for-a-generative-ai-future) and updated requirements [here](https://joss.readthedocs.io/en/latest/submitting.html#scope-and-significance). You will need to add a few sections to your paper to satisfy these new requirements.

Please let me know if you have any questions!
