### 2026-01-14 511fbac
Fix CRAN issues
