+++ paper.bib
@@ -122,3 +122,16 @@
 	url = {https://www.bioinformatics.babraham.ac.uk/projects/fastqc/},
 	urldate = {2025-04-04},
 }
+
+@article{gonatopoulos2020genetic,
+  title = {Genetic interaction mapping and exon-resolution functional genomics with a hybrid {Cas9-Cas12a} platform},
+  author = {Gonatopoulos-Pournatzis, Thomas and Aregger, Michael and Brown, Kevin R. and Farhangmehr, Shaghayegh and Braunschweig, Ulrich and Ward, Henry N. and Ha, Kevin C. H. and Weiss, Alexander and Billmann, Maximilian and Durbic, Tanja and Myers, Chad L. and Blencowe, Benjamin J. and Moffat, Jason},
+  journal = {Nature Biotechnology},
+  volume = {38},
+  number = {5},
+  pages = {638--648},
+  year = {2020},
+  publisher = {Nature Publishing Group},
+  doi = {10.1038/s41587-020-0437-z},
+  pmid = {31911676}
+}
