+++ tests/testthat/test-gi-calc.R
@@ -1,217 +1,213 @@
-skip_if_figshare_unavailable <- function() {
-  result <- get_figshare("Achilles_common_essentials.csv")
-  # Returns TRUE if Figshare is available, FALSE otherwise
-  !is.null(result)
-}
-if (skip_if_figshare_unavailable()) {
-  test_that("Test Genetic Interaction score calculations", {
-    testthat::skip_on_cran()
-    gimap_dataset <- get_example_data("gimap") %>%
-      gimap_filter() %>%
-      gimap_annotate(cell_line = "HELA") %>%
+test_that("Test Genetic Interaction score calculations", {
+  skip_if_figshare_unavailable()
+
+  gimap_dataset <- get_example_data("gimap") %>%
+    gimap_filter() %>%
+    gimap_annotate(cell_line = "HELA") %>%
+    gimap_normalize(
+      timepoints = "day",
+      missing_ids_file = tempfile()
+    ) %>%
+    calc_gi()
+
+  # Test that all expected data structures exist and are correct type
+  testthat::expect_s3_class(gimap_dataset$linear_model, "lm")
+  testthat::expect_true(is.data.frame(gimap_dataset$gi_scores))
+  testthat::expect_true(nrow(gimap_dataset$gi_scores) > 0)
+
+  # Test that required columns exist
+  expected_cols <- c("mean_expected_cs", "mean_observed_cs", "gi_score", "p_val", "fdr", "pgRNA_target", "target_type")
+  testthat::expect_true(all(expected_cols %in% colnames(gimap_dataset$gi_scores)))
+
+  # Test that values are reasonable (not NA, not infinite)
+  testthat::expect_false(any(is.na(gimap_dataset$gi_scores$mean_expected_cs)))
+  testthat::expect_false(any(is.na(gimap_dataset$gi_scores$mean_observed_cs)))
+  testthat::expect_false(any(is.na(gimap_dataset$gi_scores$gi_score)))
+  testthat::expect_false(any(is.infinite(gimap_dataset$gi_scores$gi_score)))
+
+  # Test that we have the expected target types
+  expected_target_types <- c("gene_gene", "gene_ctrl", "ctrl_gene")
+  testthat::expect_true(all(gimap_dataset$gi_scores$target_type %in% expected_target_types))
+
+  # Test that linear model has reasonable coefficients (not NA or infinite)
+  testthat::expect_false(any(is.na(gimap_dataset$linear_model$coefficients)))
+  testthat::expect_false(any(is.infinite(gimap_dataset$linear_model$coefficients)))
+
+  # Test that we have both single and double targeting results
+  testthat::expect_true("gene_gene" %in% gimap_dataset$gi_scores$target_type)
+  testthat::expect_true(any(c("gene_ctrl", "ctrl_gene") %in% gimap_dataset$gi_scores$target_type))
+})
+
+
+test_that("Test Genetic Interaction score calculations using LFC", {
+  skip_if_figshare_unavailable()
+
+  gimap_dataset <- get_example_data("gimap") %>%
+    gimap_filter() %>%
+    gimap_annotate(cell_line = "HELA") %>%
+    gimap_normalize(
+      timepoints = "day",
+      adj_method = "no_adjustment",
+      missing_ids_file = tempfile()
+    ) %>%
+    calc_gi(use_lfc = TRUE)
+
+  testthat::expect_true(class(gimap_dataset)[1] == "list")
+})
+
+
+test_that("Test Genetic Interaction score without normalization", {
+  skip_if_figshare_unavailable()
+
+  gimap_dataset_wo <- get_example_data("gimap") %>%
+    gimap_filter() %>%
[…]
