### 2026-01-14 3e1ba65
Update R/05-calculate_gi.R
Co-authored-by: marissafujimoto <fujim@uw.edu>
