+++ NEWS.md
@@ -4,6 +4,7 @@
 * Improved test skipping pattern to use proper testthat skip functions
 * All internet resource accesses fail gracefully with informative messages (CRAN policy compliance)
 * Added `"id"` to `utils::globalVariables()` to fix compatibility with dplyr 1.2.0+ (which removed the defunct `dplyr::id()`)
+* Added validation for DepMap metadata column names to handle changes in external data format gracefully
 
 # gimap 1.1.1
 
