+++ paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'gimap: An R Package for Genetic Interaction Mapping in Dual-Target CRISPR Screens'
+title: "gimap: An R Package for Genetic Interaction Mapping in Dual-Target CRISPR Screens"
 tags:
   - R
   - CRISPR
@@ -28,17 +28,17 @@ authors:
     orcid: 0000-0002-6538-2658
     affiliation: 1
 affiliations:
- - name: Fred Hutchinson Cancer Center, United States
-   index: 1
- - name: Synthesize Bio, United States
-   index: 2
+  - name: Fred Hutchinson Cancer Center, United States
+    index: 1
+  - name: Synthesize Bio, United States
+    index: 2
 date: 02 April 2025
 bibliography: paper.bib
 ---
 
 # Summary
 
-The gimap (Genetic Interaction MAPping) R package addresses a fundamental challenge in genomic research: the difficulty of understanding combinatorial interactions among genes. Gene redundancy makes traditional single-gene knockout methods ineffective for identifying therapeutic targets, as backup genes can mask the effects when a single gene is disabled. gimap offers a solution by providing a comprehensive framework for analyzing dual-target CRISPR screening data, where two genes are simultaneously disabled to reveal their backup relationships. This software implements the methods used by @parrish_discovery_2021. The package processes raw count data through a multi-step pipeline that includes normalization, calculation of expected and observed CRISPR scores, computation of genetic interaction scores, and statistical analysis to identify significant interactions. Unlike general  tools, gimap is specifically tailored for paired guide CRISPR data with built-in quality control reporting and visualization tools. The package makes best practices the default options and is available on GitHub with comprehensive documentation to support the research community in extracting meaningful insights from complex genetic screening experiments.
+The gimap (Genetic Interaction MAPping) R package addresses a fundamental challenge in genomic research: the difficulty of understanding combinatorial interactions among genes. Gene redundancy makes traditional single-gene knockout methods ineffective for identifying therapeutic targets, as backup genes can mask the effects when a single gene is disabled. gimap offers a solution by providing a comprehensive framework for analyzing dual-target CRISPR screening data, where two genes are simultaneously disabled to reveal their backup relationships. This software implements the methods used by @parrish_discovery_2021. The package processes raw count data through a multi-step pipeline that includes normalization, calculation of expected and observed CRISPR scores, computation of genetic interaction scores, and statistical analysis to identify significant interactions. Unlike general tools, gimap is specifically tailored for paired guide CRISPR data with built-in quality control reporting and visualization tools. The package makes best practices the default options and is available on GitHub with comprehensive documentation to support the research community in extracting meaningful insights from complex genetic screening experiments.
 
 # Statement of Need
 
@@ -58,7 +58,7 @@ The R package, called `gimap` (Genetic Interaction MAPping), was developed speci
 
 In order to ensure usability for the research community we built `gimap` using the following design philosophy.
 
-1. Making best practices as default options and including warning messages for when alternative options are chosen (e.g. if filtering has not been applied).  
+1. Making best practices as default options and including warning messages for when alternative options are chosen (e.g. if filtering has not been applied).
 2. Using elements from familiar packages such as fastqc reports (our `run_qc()` function creates such a report) [@fastqc].
 3. Trying to document and inform users of the statistics and decisions that have been made by the software clearly.
 
[…]
