### 2026-01-14 #8797 @fboehm
@ZeyuanSong @ByrumLab @ColemanRHarris @btmonier - would you like to review this submission for JOSS?
