### 2026-01-15 #8797 @mikerobeson
> ... would you like to review this submission for JOSS?

Thank you for the offer. Unfortunately, I'm unable to review at this time.
