### 2026-01-22 #8797 @fboehm
@anfederico @julia-wrobel @simonvandekar - would you like to review this submission for JOSS?
