### 2026-01-22 #8797 @fboehm
@buckler4 - would you like to review this submission for JOSS?
