### 2026-01-23 #9891 @editorialbot
[REVIEW]: gimap: An R Package for Genetic Interaction Mapping in Dual-Target CRISPR Screens
**Submitting author:** @cansavvy (<a href="http://orcid.org/0000-0001-6331-7070">Candace Savonen</a>)
**Repository:** https://github.com/FredHutch/gimap
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.1
