### 2026-02-03 #9891 @btmonier
Hi @cansavvy - this looks like a useful package for the CRISPR community. The software is clear, easy to install and to use. Some minor comments I have before I can check the last boxes in my checklist:

**Licensing Conflicts** - In your repository, you state [CC0-1.0](https://github.com/FredHutch/gimap/blob/main/LICENSE) (which is not [OSI approved](https://opensource.org/licenses/alphabetical)), but in your [`DESCRIPTION`](https://github.com/FredHutch/gimap/blob/2013259ab18d3b5237bc76fb67c0a6c1eafb3e55/DESCRIPTION#L18) file, you state GPL-3.

**Installation** - In your README, you specify installation via the `remotes` package, but on the [provided website](https://fredhutch.github.io/gimap/articles/quick-start.html#requirements), you also mention installation through CRAN using `install.packages("gimap")`. I was unable to install it with the CRAN approach since it appears to be missing:

```
Warning message:
package ‘gimap’ is not available for this version of R

A version of this package for your version of R might be available elsewhere,
see the ideas at
https://cran.r-project.org/doc/manuals/r-patched/R-admin.html#Installing-packages
```

**Community Guidelines** - please set up issue tracking on this repo in case the community wants to reach out for possible bugs, feature requests, enhancements, etc.

**Paper Submission** - Some nitpicks:

* In [`paper.bib`](https://github.com/FredHutch/gimap/blob/2013259ab18d3b5237bc76fb67c0a6c1eafb3e55/paper.bib#L13) you specify an entry that is not referenced in the paper (i.e. the articled tagged as `bioinformatics`)
* In the [`paper.md` front matter](https://github.com/FredHutch/gimap/blob/2013259ab18d3b5237bc76fb67c0a6c1eafb3e55/paper.md?plain=1#L20), you list one of the authors with the last name of "Groso", but in the [`DESCRIPTION`](https://github.com/FredHutch/gimap/blob/2013259ab18d3b5237bc76fb67c0a6c1eafb3e55/DESCRIPTION#L14) file, you specify this individual's last name as "Grosso"
* The paper detailing the "Benjamini-Hochberg" procedure (tagged as `benjamini1995controlling` in your [`paper.bib`](https://github.com/FredHutch/gimap/blob/2013259ab18d3b5237bc76fb67c0a6c1eafb3e55/paper.bib#L2)) is missing a DOI (as flagged by the bot) - you should add some `doi` and `url` keys to this BibTeX entry, e.g.:

  ``` bib
  @article{benjamini1995controlling,
    title = {Controlling the false discovery rate: a practical and powerful approach to multiple testing},
    author = {Benjamini, Yoav and Hochberg, Yosef},
    journal = {Journal of the Royal Statistical Society: Series B (Methodological)},
    volume = {57},
    number = {1},
    pages = {289--300},
    year = {1995},
    publisher = {Wiley Online Library},
    doi = {10.1111/j.2517-6161.1995.tb02031.x},
    url = {https://doi.org/10.1111/j.2517-6161.1995.tb02031.x}
  }
  ```

**Example Usage** - Looking at this code in the [quick start guide](https://fredhutch.github.io/gimap/articles/quick-start.html#setting-up-data):

  ``` r
  gimap_dataset <- gimap_dataset %>%
      gimap_filter() %>%
      gimap_annotate(cell_line = "HELA") %>%
      gimap_normalize(
          timepoints = "day"
      ) %>%
      calc_gi()
  ```

It seems that something has changed with the returned data and is causing downstream exceptions which needs to be addressed:

```
DepMap metadata format has changed - expected column 'stripped_cell_line_name' not found. Returning dataset without cell line annotation.
Error in gimap_normalize(., timepoints = "day") :
  No annotations are stored in this gimap_dataset, annotations are neededso we know what genes should be used as controls.Please run gimap_annotate() function on your gimap_dataset and then retry this function.
4.
stop("No annotations are stored in this gimap_dataset, annotations are needed",
"so we know what genes should be used as controls.", "Please run gimap_annotate() function on your gimap_dataset and then",
" retry this function.")
3.
gimap_normalize(., timepoints = "day")
2.
calc_gi(.)
1.
gimap_dataset %>% gimap_filter() %>% gimap_annotate(cell_line = "HELA") %>%
gimap_normalize(timepoints = "day") %>% calc_gi()
```

**Site builder** - consider adding a GitHub action to build the site on triggers (e.g. PRs to main, etc) using `pkgdown`. You currently build the site using pkgdown but store all of the HTML/JS intermediates on the main branch which will increase the payload size of this package.
