+++ tests/testthat/test-annotate.R
@@ -1,8 +1,9 @@
 test_that("Annotation options", {
   skip_if_figshare_unavailable()
   skip_if_depmap_changed()
+  data_dir <- test_example_data_dir()
 
-  gimap_dataset <- get_example_data("gimap") %>%
+  gimap_dataset <- get_example_data("gimap", data_dir = data_dir) %>%
     gimap_filter() %>%
     gimap_annotate(cell_line = "HELA")
 
@@ -18,13 +19,13 @@ test_that("Annotation options", {
   # It should warn you if you try to say FALSE for cell line_annotate but
   # don't provide a custom_tpm or use normalize_by_unexpressed = FALSE
   testthat::expect_error(
-    gimap_dataset <- get_example_data("gimap") %>%
+    gimap_dataset <- get_example_data("gimap", data_dir = data_dir) %>%
       gimap_filter() %>%
       gimap_annotate(cell_line_annotate = FALSE) %>%
       gimap_normalize(timepoints = "day")
   )
 
-  gimap_dataset <- get_example_data("gimap") %>%
+  gimap_dataset <- get_example_data("gimap", data_dir = data_dir) %>%
     gimap_filter() %>%
     gimap_annotate(cell_line_annotate = FALSE) %>%
     gimap_normalize(
@@ -37,17 +38,15 @@ test_that("Annotation options", {
     colnames(gimap_dataset$normalized_log_fc)))
 
   ## Try out using custom TPM data
+  tpm_file <- tpm_setup(data_dir = data_dir)
   custom_tpm <- vroom::vroom(
-    file.path(
-      system.file("extdata", package = "gimap"),
-      "CCLE_expression.csv"
-    ),
+    tpm_file,
     show_col_types = FALSE,
     col_select = c("genes", "ACH-001086")
   ) %>%
     dplyr::rename(log2_tpm = `ACH-001086`)
 
-  gimap_dataset <- get_example_data("gimap") %>%
+  gimap_dataset <- get_example_data("gimap", data_dir = data_dir) %>%
     gimap_filter() %>%
     gimap_annotate(
       custom_tpm = custom_tpm,
