+++ tests/testthat/test-normalize.R
@@ -1,8 +1,9 @@
 test_that("Test normalization", {
   skip_if_figshare_unavailable()
   skip_if_depmap_changed()
+  data_dir <- test_example_data_dir()
 
-  gimap_dataset <- get_example_data("gimap") %>%
+  gimap_dataset <- get_example_data("gimap", data_dir = data_dir) %>%
     gimap_filter() %>%
     gimap_annotate(cell_line = "HELA") %>%
     gimap_normalize(
@@ -41,8 +42,9 @@ test_that("Test normalization", {
 test_that("Test normalization without expression cutoff", {
   skip_if_figshare_unavailable()
   skip_if_depmap_changed()
+  data_dir <- test_example_data_dir()
 
-  gimap_dataset_true <- get_example_data("gimap") %>%
+  gimap_dataset_true <- get_example_data("gimap", data_dir = data_dir) %>%
     gimap_filter() %>%
     gimap_annotate(cell_line = "HELA") %>%
     gimap_normalize(
@@ -51,7 +53,7 @@ test_that("Test normalization without expression cutoff", {
       missing_ids_file = tempfile()
     )
 
-  gimap_dataset_false <- get_example_data("gimap") %>%
+  gimap_dataset_false <- get_example_data("gimap", data_dir = data_dir) %>%
     gimap_filter() %>%
     gimap_annotate(cell_line = "HELA") %>%
     gimap_normalize(
