+++ R/utils.R
@@ -43,13 +43,49 @@ check_internet_available <- function(url = "https://api.figshare.com",
   )
 }
 
+#' Build manual download instructions for Figshare files
+#' @description Internal helper to build a warning message for manual downloads.
+#' @param item Figshare item ID.
+#' @param files Character vector of file names.
+#' @return A formatted message string.
+#' @keywords internal
+figshare_manual_download_message <- function(item, files) {
+  files_text <- paste0(files, collapse = ", ")
+  paste0(
+    "You can manually download these files from Figshare item ", item, ": ",
+    files_text, ". ",
+    "Use `gimap::get_figshare(file_name = ..., item = \"", item,
+    "\", output_dir = <data_dir>)` or see README \"Manual data download (Figshare)\"."
+  )
+}
+
+#' Resolve and create data directory
+#' @description Internal helper to resolve a writable data directory for downloads.
+#' @param data_dir Optional data directory. If NULL or empty, uses a user data dir.
+#' @return A normalized, existing directory path.
+#' @keywords internal
+gimap_data_dir <- function(data_dir = NULL) {
+  if (is.null(data_dir) || length(data_dir) == 0) {
+    data_dir <- tools::R_user_dir("gimap", "data")
+  } else if (is.na(data_dir) || !nzchar(data_dir)) {
+    data_dir <- tools::R_user_dir("gimap", "data")
+  }
+
+  data_dir <- normalizePath(data_dir, winslash = "/", mustWork = FALSE)
+  if (!dir.exists(data_dir)) {
+    dir.create(data_dir, recursive = TRUE, showWarnings = FALSE)
+  }
+  return(data_dir)
+}
+
 
 #' Returns example data for package
 #' @description This function loads and returns example data for the package.
 #' Which dataset is returned must be specified. Data will be downloaded from Figshare
 #' the first time it is used.
 #' @param which_data options are "count" or "meta"; specifies which example dataset should be returned
-#' @param data_dir Where should the data be saved if applicable?
+#' @param data_dir Where should the data be saved if applicable? If NULL, a user data
+#' directory will be used.
 #' @param refresh_data should the example data that's been downloaded be deleted
 #' and redownloaded?
 #' @export
@@ -66,8 +102,9 @@ check_internet_available <- function(url = "https://api.figshare.com",
 #' annotation <- get_example_data("annotation")
 #' }
 get_example_data <- function(which_data,
-                             data_dir = system.file("extdata", package = "gimap"),
+                             data_dir = NULL,
                              refresh_data = FALSE) {
+  data_dir <- gimap_data_dir(data_dir)
   file_name <- switch(which_data,
     "count" = "PP_pgPEN_HeLa_counts.txt",
     "count_treatment" = "counts_pgPEN_PC9_example.tsv",
@@ -84,12 +121,12 @@ get_example_data <- function(which_data,
 
   file_path <- file.path(data_dir, file_name)
 
-  if (!grepl("RDS$", file_name)) {
-    # Save file path in the options
-    file_path_list <- list(file_path)
-    names(file_path_list) <- which_data
-    options(file_path_list)
+  # Save file path in the options
+  file_path_list <- list(file_path)
+  names(file_path_list) <- which_data
+  options(file_path_list)
 
+  if (!grepl("RDS$", file_name)) {
     if (!file.exists(file_path)) {
       download_result <- get_figshare(
         file_name = file_name,
@@ -98,9 +135,10 @@ get_example_data <- function(which_data,
       )
       # Handle case where download failed
       if (is.null(download_result)) {
-        message(
+        warning(
[…]
