+++ tests/testthat/test-setup_data.R
@@ -6,6 +6,7 @@ example_sample_metadata <- data.frame(id = 1:5, replicate = factor(c(1, 1, 2, 2,
 # Test elements inside output list
 test_that("setup_data() works correctly", {
   skip_if_figshare_unavailable()
+  data_dir <- test_example_data_dir()
 
   result <- setup_data(
     counts = example_counts,
@@ -20,10 +21,10 @@ test_that("setup_data() works correctly", {
   expect_equal(result$counts_per_sample, apply(example_counts, 2, sum))
 
 
-  counts_timepoint <- get_example_data("count")
-  counts_treatment <- get_example_data("count_treatment")
-  gimap_dataset <- get_example_data("gimap")
-  gimap_dataset <- get_example_data("gimap_treatment")
-  metadata <- get_example_data("meta")
-  annotation <- get_example_data("annotation")
+  counts_timepoint <- get_example_data("count", data_dir = data_dir)
+  counts_treatment <- get_example_data("count_treatment", data_dir = data_dir)
+  gimap_dataset <- get_example_data("gimap", data_dir = data_dir)
+  gimap_dataset <- get_example_data("gimap_treatment", data_dir = data_dir)
+  metadata <- get_example_data("meta", data_dir = data_dir)
+  annotation <- get_example_data("annotation", data_dir = data_dir)
 })
