+++ man/save_example_timepoint_data.Rd
@@ -4,7 +4,10 @@
 \alias{save_example_timepoint_data}
 \title{Set up example count data}
 \usage{
-save_example_timepoint_data()
+save_example_timepoint_data(data_dir = NULL)
+}
+\arguments{
+\item{data_dir}{Optional data directory override.}
 }
 \value{
 Returns the file path to folder where the example data is stored

[…]
