### 2026-02-18 #9891 @fboehm
@tkchafin @tomsing1 @ColemanRHarris @ZeyuanSong @preetida @rspirgel @shahmoradi @MohmedSoudy - we need another reviewer for this submission. Would you like to review it for JOSS?

Thanks!
