### 2026-02-20 #9891 @kalyanidhusia
Thank you for the opportunity to review this submission.

I apologize for the delay and any disruption this may cause to the editorial process. Hope I'm not too late @fboehm
Below is my structured review checklist following the [JOSS reviewer guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html)

---

## Conflict of Interest & Ethics

- [x] **Conflict of Interest:** I confirm that I have read the [JOSS COI policy](https://joss.readthedocs.io/en/latest/policies.html#conflict-of-interest) and have no associations with the authors (collaborations in the last 4 years, same institution in the last year, or lifetime mentor/student relationships) that would prevent an impartial judgment.
- [x] **Code of Conduct:** I agree to abide by the [JOSS Code of Conduct](https://joss.theoj.org/about#code-of-conduct), ensuring all interactions with the authors are professional and constructive.
- [x] **AI Usage:** I confirm that I have read the [JOSS AI policy](https://joss.readthedocs.io/en/latest/policies.html#ai-tools-policy). I will perform all evaluative assessments independently and disclose any non-substantive AI assistance (e.g., language and grammar refinement).

---

- [x] **Repository:** https://github.com/FredHutch/gimap
- [x] **License:** https://github.com/FredHutch/gimap/blob/main/LICENSE
- [x] **Tests directory:** https://github.com/FredHutch/gimap/tree/main/tests/testthat
- [x] **Software paper (`paper.md`):** https://github.com/FredHutch/gimap/blob/main/paper.md
- [x] **Substantial Scholarly Effort:** The software represents a significant research contribution (not a minor utility or simple wrapper).
- [ ] **Authorship & Contribution:** The author list appears complete and the submitting author has made a major contribution.

---

- [ ] **Installation:** Installation proceeds as outlined in the README:
      https://github.com/FredHutch/gimap
- [ ] **Example usage / vignette:**
      https://fredhutch.github.io/gimap/articles/quick-start.html
- [ ] **Functional Claims:** The functional claims of the software have been independently confirmed.
- [ ] **Performance Claims:** Any stated performance claims have been independently verified.

### Normalization Assessment

- [ ] **Normalization Methodology:** The statistical and computational rationale for normalization is clearly defined (e.g., library size scaling, quantile normalization, variance-stabilizing transformation).
- [ ] **Suitability:** The normalization technique is appropriate for dual-target CRISPR screen data (handling sparsity and zero inflation).
- [ ] **Transparency:** The documentation/API allows inspection of normalization effects (e.g., diagnostic plots such as MA/MD plots or distribution plots).
- [ ] **Flexibility:** Users can select between normalization methods, or the implementation clearly justifies a fixed approach.

---

- [ ] **Statement of Need:** The target audience is clearly defined, and the problem the software solves is well-articulated.
- [ ] **Installation Instructions:** Clear, step-by-step installation guidance including dependencies.
- [ ] **Example Dataset / Workflow:** A realistic example workflow (e.g., CRISPR screen analysis).
- [ ] **API Documentation:** Core functionality (functions, classes, methods) is documented sufficiently to understand inputs and outputs.
- [ ] **Community Guidelines:** Clear instructions on how to:
  - [ ] Contribute to the software
  - [ ] Report issues or bugs
  - [ ] Seek support

---

## Software Paper (`paper.md`)

- [x] **Summary:** Clear description suitable for a broad scientific audience.
- [x] **Statement of Need:** The paper explains why the software was developed and its significance.
- [ ] **State of the Field:** The manuscript compares *gimap* to related tools (e.g., CRISPR analysis or matrix integration packages).
- [ ] **Supplementary Materials (if applicable):** Any provided demo or instructional video appropriately supports the documentation and workflow description.
- [x] **References:** References are complete, include DOIs where available, and use proper citation syntax.
