### 2026-02-20 #9891 @tomsing1
- No, the current license is not OSI approved [issue raised](https://github.com/FredHutch/gimap/issues/109)

  - The R package makes a potentially destructive call by deleting files whose location is specified in undocumented options (such as `options("counts")` or `options("meta")`). [This issue](https://github.com/FredHutch/gimap/issues/111) needs to be mitigated, at minimum with better documentation.
  - As previously reported by @btmonier, the Quick Start vignette doesn't run successfully, [raised issue here](https://github.com/FredHutch/gimap/issues/112).

  - The [Quick Start vignette]() could be improved, at a minimum two important hidden code cells should be shown. [Issue raised](https://github.com/FredHutch/gimap/issues/114)
  - Yes, but the error reported in [this issue](https://github.com/FredHutch/gimap/issues/112) causes 8 out of the 9 unit tests to be skipped.
