### 2026-02-22 FredHutch/gimap#109 @tomsing1
[JOSS review] OSI approved open source license
This question is part of the [General checks](https://joss.readthedocs.io/en/latest/review_checklist.html#general-checks) of the JOSS review:
- Does the repository contain a plain-text LICENSE file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license?

The [LICENSE](https://github.com/FredHutch/gimap/blob/main/LICENSE) file included in this repository includes the Creative Commons Zero v1.0 Universal license. It does not seem to be on the list of the OSI approved licenses. Please consider changing the license to one that is approved?
