### 2026-02-22 FredHutch/gimap#111 @tomsing1
[JOSS review] Undocumented options and unsafe use of unlink()
The `delete_example_data()` function issues an uninformative message `"Deleting the example data files listed in options"`, but the user doesn't know which options that refers to.

Debugging the function reveals that it is looking up the following options: `count`, `count_treatment`, `meta`, `gimap`, `gimap_treatment`, and `annotation`.

The `delete_example_data()` then deletes whichever filename may or may not be stored in these options. That's very dangerous, as the names of these options is very generic[^1]  and their use is not documented anywhere.

[^1]: Check e.g. how the {renv} package names and documents the options it relies on [here](https://rstudio.github.io/renv/reference/config.html).

Issuing a destructive call such as `unlink` must have better safeguards. (E.g. what if a user unknowingly sets their own `counts` option - and points to an important file somewhere unrelated to {gimap}?)

Currently, these files are stored in the `gimap_data_dir`, which is application specific [^2]- and deterministic. If you want to cache example files on the user's disk, **don't** enable users to change this location. (You are already returning a message where they are saved, users can always copy them to a different location if they like.)

By controlling the location of the example files, you can construct the file paths that should be deleted (e.g. specific filenames within the application-specific directory). That way, you can make sure your deletion function is _only_ ever deleting files from the {gimap} application-specific folder. No need to use `options` for that, which you don't have control over.

[^2]: If you haven't encountered the [appdirs::user_data_dir](https://rappdirs.r-lib.org/reference/user_data_dir.html) function, yet, then perhaps that could be useful for future projects.

Finally, the [complete set of gimap example files is relatively small](https://figshare.com/articles/dataset/gimap_example_data/28264271). You might want to consider _not_ offering to store them beyond the life of the current R session (e.g. by always placing the downloaded file into the `tempdir()` - and then retrieving them again next time. That's the option I would choose: reduce the complexity of your codebase and don't take on responsibility for adding / deleting files from the user's hard-drive (outside of the tempdir).
