### 2026-02-22 FredHutch/gimap#112 @tomsing1
[JOSS review] Quick Start example raises an error
I am following the [Quick Start vignette](https://fredhutch.github.io/gimap/articles/quick-start.html), but unfortunately I am running into an error:

```r
library(gimap)
library(dplyr)

# start with a clean slate
delete_example_data()

output_dir <- "output_timepoints"
dir.create(output_dir, showWarnings = FALSE)

example_data <- get_example_data("count")

counts <- example_data %>%
  select(c("Day00_RepA", "Day22_RepA", "Day22_RepB", "Day22_RepC")) %>%
  as.matrix()
pg_ids <- example_data %>%
  dplyr::select("id")
sample_metadata <- data.frame(
  col_names = c("Day00_RepA", "Day22_RepA", "Day22_RepB", "Day22_RepC"),
  day = as.numeric(c("0", "22", "22", "22")),
  rep = as.factor(c("RepA", "RepA", "RepB", "RepC"))
)
gimap_dataset <- setup_data(
  counts = counts,
  pg_ids = pg_ids,
  sample_metadata = sample_metadata
)
gimap_dataset <- gimap_dataset %>%
  gimap_filter() %>%
  gimap_annotate(cell_line = "HELA") %>%
  gimap_normalize(
    timepoints = "day"
  ) %>%
  calc_gi()
```

raises:

```
DepMap metadata format has changed - expected column 'stripped_cell_line_name' not found. Returning dataset without cell line annotation.
Error in gimap_normalize(., timepoints = "day") :
  No annotations are stored in this gimap_dataset, annotations are neededso we know what genes should be used as controls.Please run gimap_annotate() function on your gimap_dataset and then retry this function.
```

<summary>
Session Information
</summary>

```
R version 4.5.2 (2025-10-31)
Platform: aarch64-apple-darwin20
Running under: macOS Tahoe 26.3

Matrix products: default
BLAS:   /System/Library/Frameworks/Accelerate.framework/Versions/A/Frameworks/vecLib.framework/Versions/A/libBLAS.dylib
LAPACK: /Library/Frameworks/R.framework/Versions/4.5-arm64/Resources/lib/libRlapack.dylib;  LAPACK version 3.12.1

locale:
[1] en_US.UTF-8/en_US.UTF-8/en_US.UTF-8/C/en_US.UTF-8/en_US.UTF-8

time zone: America/Los_Angeles
tzcode source: internal

attached base packages:
[1] stats     graphics  grDevices datasets  utils     methods   base

other attached packages:
[1] dplyr_1.2.0 gimap_1.1.2

 [1] bit_4.6.0          gtable_0.3.6       jsonlite_2.0.0     crayon_1.5.3
 [5] compiler_4.5.2     renv_1.1.7         tidyselect_1.2.1   stringr_1.6.0
 [9] parallel_4.5.2     snakecase_0.11.1   tidyr_1.3.2        scales_1.4.0
[13] yaml_2.3.12        ggplot2_4.0.2      readr_2.2.0        R6_2.6.1
[17] generics_0.1.4     curl_7.0.0         tibble_3.3.1       janitor_2.2.1
[21] openssl_2.3.4      lubridate_1.9.5    tzdb_0.5.0         pillar_1.11.1
[25] RColorBrewer_1.1-3 rlang_1.1.7        stringi_1.8.7      S7_0.2.1
[29] bit64_4.6.0-1      timechange_0.4.0   cli_3.6.5          withr_3.0.2
[33] magrittr_2.0.4     grid_4.5.2         vroom_1.7.0        hms_1.1.4
[37] askpass_1.2.1      lifecycle_1.0.5    vctrs_0.7.1        pheatmap_1.0.13
[41] glue_1.8.0         farver_2.1.2       purrr_1.2.1        httr_1.4.8
[45] tools_4.5.2        pkgconfig_2.0.3
```
