### 2026-02-22 FredHutch/gimap#112 @tomsing1
The `DepMap metadata format has changed` message also leads to 8 out of the 17 unit tests to be skipped:

```
══ Skipped tests (8) ═══════════════════════════════════════════════════════════
• DepMap metadata format has changed or is unavailable (8):
  'test-annotate.R:3:3', 'test-gi-calc.R:3:3', 'test-gi-calc.R:46:3',
  'test-gi-calc.R:65:3', 'test-gi-calc.R:112:3', 'test-gi-calc.R:187:3',
  'test-normalize.R:3:3', 'test-normalize.R:44:3'

[ FAIL 0 | WARN 0 | SKIP 8 | PASS 9 ]
```
