### 2026-02-22 FredHutch/gimap#114 @tomsing1
[JOSS review] Quick Start vignette improvements
Suggestions for improving the [Quick Start vignette](https://fredhutch.github.io/gimap/articles/quick-start.html):

- The [two code cells](https://github.com/FredHutch/gimap/blob/781cfff4da1b7c3f9fa87130cb65f60a6724833a/vignettes/quick-start.Rmd#L44) that attach the {gimap} and {dplyr} package are hidden in the rendered HTML file - but they are an essential step for users to run the following code successfully. Perhaps you wanted to suppress startup messages with `suppressPackageStartupMessages()`?
- To make the easier to follow and help users map their own experiments onto the workflow, consider including a description of the study design, and a link to the original study from which the example data is taken (so users can read up on the details).
- The `Setup` section of the vignette should explain what the `example_data `object is, e.g.
  > The object is a tibble with the following columns ....
- The `calc_gi()` function is called without any further explanation, followed by Section [Example output](https://fredhutch.github.io/gimap/articles/quick-start.html#example-output)  that starts with "Genetic interaction is calculated by:" It then lists different different terms in a bulleted list but these are not really explained. This section might be a stub?
- Now that the example data is hosted on figshare, it would be nice to include a link to the [figshare HTML page](https://figshare.com/articles/dataset/gimap_example_data/28264271), not just the item identifier, in the [Manual data download (Figshare)](https://github.com/FredHutch/gimap/blob/main/README.md#example-data-figshare-item-28264271) section.
