### 2026-02-22 #9891 @tomsing1
> Thanks, [@tomsing1](https://github.com/tomsing1) and [@kalyanidhusia](https://github.com/kalyanidhusia) for creating the checklists and starting the reviews. Please let me know how the reviews are going as you get started.

I have raised a few issues in the the [target repository](https://github.com/FredHutch/gimap). Most of them a relatively minor. But the package is currently not functional, as pointed out by @btmonier a few weeks ago. [That issue](https://github.com/FredHutch/gimap/issues/112) seems to be a blocker that needs to be addressed by the authors before we can e.g. run all of the example code and reproduce the package's functionality. (The package only passes its tests because the authors included a function that skips 8/17 unit tests precisely when the current issue arises. So this might have been a known issue 🤔 ?)
