### 2026-02-24 FredHutch/gimap#112 @cansavvy
> The `DepMap metadata format has changed` message also leads to 8 out of the 17 unit tests to be skipped:
>
> ```
> ══ Skipped tests (8) ═══════════════════════════════════════════════════════════
> • DepMap metadata format has changed or is unavailable (8):
>   'test-annotate.R:3:3', 'test-gi-calc.R:3:3', 'test-gi-calc.R:46:3',
>   'test-gi-calc.R:65:3', 'test-gi-calc.R:112:3', 'test-gi-calc.R:187:3',
>   'test-normalize.R:3:3', 'test-normalize.R:44:3'
>
> [ FAIL 0 | WARN 0 | SKIP 8 | PASS 9 ]
> ```

Thanks for this report @tomsing1

This is due to a CRAN restriction that's putting me in a bit of bind.

It doesn't want the package to be big (so I can't keep even basic annotation files already in there).

But on CRAN they also don't want you to download extra files for use either.

So I had to turn off unit tests for remote testing, but the tests were passing locally.
