### 2026-04-06 #9891 @fboehm
@Kevin-Mattheus-Moerman - What is the procedure for rejecting a JOSS submission? The submitting author has not replied to recent queries. Thanks.
