### 2026-04-06 #9891 @cansavvy
> @Kevin-Mattheus-Moerman - What is the procedure for rejecting a JOSS submission? The submitting author has not replied to recent queries. Thanks.
>

So sorry for the delay. I have switched institutions and roles and am no longer paid to work on this so it's taking me awhile to find free time
