### 2026-04-16 FredHutch/gimap#109 @cansavvy
Addressed in https://github.com/FredHutch/gimap/commit/215156efe4c86b222bea4335f604cf8554885ead
