### 2026-04-16 91f5dd4
Add CRAN instructions to readme
