+++ README.md
@@ -5,18 +5,24 @@
 <!-- DON'T EDIT THIS SECTION, INSTEAD RE-RUN doctoc TO UPDATE -->
 **Table of Contents**  *generated with [DocToc](https://github.com/thlorenz/doctoc)*
 
-  - [Background on paired guide CRISPR](#background-on-paired-guide-crispr)
-  - [What is gimap?](#what-is-gimap)
-  - [Example output from this package:](#example-output-from-this-package)
-  - [pgPEN library design](#pgpen-library-design)
-  - [About Genetic Interaction Scores](#about-genetic-interaction-scores)
-      - [Expected CRISPR scores](#expected-crispr-scores)
-    - [Normalization](#normalization)
-  - [Prerequisites](#prerequisites)
-  - [Manual data download (Figshare)](#manual-data-download-figshare)
-  - [Getting Started Tutorial](#getting-started-tutorial)
-  - [Docker image](#docker-image)
-  - [Citations:](#citations)
+- [Background on paired guide CRISPR](#background-on-paired-guide-crispr)
+- [What is gimap?](#what-is-gimap)
+- [Example output from this package:](#example-output-from-this-package)
+- [pgPEN library design](#pgpen-library-design)
+- [About Genetic Interaction Scores](#about-genetic-interaction-scores)
+    - [Expected CRISPR scores](#expected-crispr-scores)
+  - [t-tests](#t-tests)
+  - [Normalization](#normalization)
+- [Prerequisites](#prerequisites)
+  - [Release version (CRAN)](#release-version-cran)
+  - [Development version (GitHub)](#development-version-github)
+- [Manual data download (Figshare)](#manual-data-download-figshare)
+  - [Where to put files](#where-to-put-files)
+  - [Example data (Figshare item 28264271)](#example-data-figshare-item-28264271)
+  - [DepMap/CCLE annotation data (Figshare item 19700056)](#depmapccle-annotation-data-figshare-item-19700056)
+- [Getting Started Tutorial](#getting-started-tutorial)
+- [Docker image](#docker-image)
+- [Citations:](#citations)
 
 <!-- END doctoc generated TOC please keep comment here to allow auto update -->
 
@@ -163,11 +169,26 @@ log2FC adjusted =
 
 ## Prerequisites
 
-In order to run this pipeline you will need R and to install the `gimap` package and its dependencies. In R you can run this to install the package:
+In order to run this pipeline you will need R and to install the `gimap` package and its dependencies.
+
+### Release version (CRAN)
+
+The [latest release on CRAN](https://cran.r-project.org/package=gimap) is the recommended install for most users:
+
+```
+install.packages("gimap")
+```
+
+### Development version (GitHub)
+
+The `main` branch on GitHub may contain changes that are not yet on CRAN. If you need the development build, install with the [remotes](https://cran.r-project.org/package=remotes) package:
+
 ```
 install.packages("remotes")
 remotes::install_github("FredHutch/gimap")
 ```
+
+To install from a specific branch or tag, pass the `ref` argument (for example `ref = "main"` or a release tag from the [releases page](https://github.com/FredHutch/gimap/releases)).
 
 ## Manual data download (Figshare)
 
