### 2026-04-16 94ee575
Address unlink problem and unnecessary directory creation
