+++ R/utils.R
@@ -78,16 +78,75 @@ gimap_data_dir <- function(data_dir = NULL) {
   return(data_dir)
 }
 
+#' Session cache for Figshare example files (under [tempdir()])
+#' @description Not exported; avoids persisting example downloads outside the session
+#' temp directory unless the caller passes an explicit `data_dir`.
+#' @return Normalized path to an existing directory.
+#' @keywords internal
+.gimap_example_cache_dir <- function() {
+  d <- file.path(tempdir(), "gimap-example-data")
+  d <- normalizePath(d, winslash = "/", mustWork = FALSE)
+  if (!dir.exists(d)) {
+    dir.create(d, recursive = TRUE, showWarnings = FALSE)
+  }
+  normalizePath(d, winslash = "/", mustWork = TRUE)
+}
+
+#' Default directory for [get_example_data()] and related helpers
+#' @keywords internal
+resolve_example_data_dir <- function(data_dir) {
+  if (is.null(data_dir) || length(data_dir) == 0L ||
+    (length(data_dir) == 1L && (is.na(data_dir) || !nzchar(data_dir)))) {
+    .gimap_example_cache_dir()
+  } else {
+    gimap_data_dir(data_dir)
+  }
+}
+
+#' Basenames of all files managed by [get_example_data()]
+#' @keywords internal
+.gimap_example_filenames <- function() {
+  c(
+    "PP_pgPEN_HeLa_counts.txt",
+    "counts_pgPEN_PC9_example.tsv",
+    "pgRNA_ID_pgPEN_library_comp.csv",
+    "gimap_dataset_timepoint.RDS",
+    "gimap_dataset_treatment.RDS",
+    "pgPEN_annotations.txt"
+  )
+}
+
+#' Whether `dir` is allowed as a root for [delete_example_data()]
+#' @keywords internal
+.is_safe_example_data_delete_root <- function(dir) {
+  dir <- normalizePath(dir, winslash = "/", mustWork = FALSE)
+  legacy <- normalizePath(
+    tools::R_user_dir("gimap", "data"),
+    winslash = "/", mustWork = FALSE
+  )
+  td <- normalizePath(tempdir(), winslash = "/", mustWork = FALSE)
+  cache <- normalizePath(
+    file.path(tempdir(), "gimap-example-data"),
+    winslash = "/", mustWork = FALSE
+  )
+  dir == legacy ||
+    dir == cache ||
+    startsWith(dir, paste0(td, .Platform$file.sep))
+}
+
 
 #' Returns example data for package
 #' @description This function loads and returns example data for the package.
 #' Which dataset is returned must be specified. Data will be downloaded from Figshare
 #' the first time it is used.
-#' @param which_data options are "count" or "meta"; specifies which example dataset should be returned
-#' @param data_dir Where should the data be saved if applicable? If NULL, a user data
-#' directory will be used.
-#' @param refresh_data should the example data that's been downloaded be deleted
-#' and redownloaded?
+#' @param which_data One of `"count"`, `"count_treatment"`, `"meta"`, `"gimap"`,
+#' `"gimap_treatment"`, or `"annotation"`; specifies which example dataset is returned.
+#' @param data_dir Where files are stored. If `NULL` (default), example files are
+#' written under `file.path(tempdir(), "gimap-example-data")` for the current R
+#' session (they are not kept after the session ends). Pass an explicit directory
+#' if you want a persistent copy (for example to mirror files elsewhere).
+#' @param refresh_data If `TRUE`, removes known example filenames from `data_dir`
+#' (only under safe locations; see [delete_example_data()]) before re-downloading.
 #' @export
 #' @returns the respective example data either as a data frame or a specialized
 #' gimap_dataset depending on what was requested.
[…]
