### 2026-04-16 FredHutch/gimap#111 @cansavvy
Addressed in: https://github.com/FredHutch/gimap/commit/94ee575e483981d4793dc3b7711006a43b723240

We now have a safer delete_example_data() — It only removes fixed basenames from .gimap_example_filenames(), and only when the directory is one of:

- that session example cache under tempdir(),
- any subdirectory of tempdir() (so tests and tempfile() dirs still work),
or the legacy tools::R_user_dir("gimap", "data") so old installs can be cleaned up.
- Otherwise it stop()s with a clear message. The message lists each file removed.
