+++ R/03-annotate.R
@@ -1,23 +1,43 @@
 #' Normalize DepMap metadata for downstream joins
 #' @keywords internal
 normalize_depmap_metadata <- function(depmap_metadata) {
-  # Accept alternative column names used in different DepMap/Figshare releases
+  if (is.null(depmap_metadata) || nrow(depmap_metadata) < 1L) {
+    return(NULL)
+  }
   cols <- colnames(depmap_metadata)
-  # Standardize DepMap ID column (e.g. depmap_id -> DepMap_ID)
+  # Strip UTF-8 BOM from the first column name if present (bad CSV export / download)
+  if (length(cols) > 0L) {
+    cols[1] <- sub("^\ufeff", "", cols[1])
+    colnames(depmap_metadata) <- cols
+    cols <- colnames(depmap_metadata)
+  }
+  # Standardize DepMap ID column across releases (Figshare / portal naming)
   if (!("DepMap_ID" %in% cols) && ("depmap_id" %in% cols)) {
     depmap_metadata <- depmap_metadata %>%
       dplyr::rename(DepMap_ID = depmap_id)
   }
+  cols <- colnames(depmap_metadata)
+  if (!("DepMap_ID" %in% cols) && ("ModelID" %in% cols)) {
+    depmap_metadata <- depmap_metadata %>%
+      dplyr::rename(DepMap_ID = ModelID)
+  }
+  cols <- colnames(depmap_metadata)
+  if (!("DepMap_ID" %in% cols) && ("model_id" %in% cols)) {
+    depmap_metadata <- depmap_metadata %>%
+      dplyr::rename(DepMap_ID = model_id)
+  }
+  cols <- colnames(depmap_metadata)
   # Standardize stripped cell line name column
   if (!("stripped_cell_line_name" %in% cols)) {
     if ("cell_line" %in% cols) {
-      # Derive stripped name: part before first underscore, uppercased
       depmap_metadata <- depmap_metadata %>%
         dplyr::mutate(stripped_cell_line_name = toupper(gsub("_.*$", "", .data$cell_line)))
     } else if ("cell_line_name" %in% cols) {
-      # Use cell_line_name with non-alphanumeric stripped
       depmap_metadata <- depmap_metadata %>%
         dplyr::mutate(stripped_cell_line_name = toupper(gsub("[^A-Za-z0-9]", "", .data$cell_line_name)))
+    } else if ("Cell_Line_Name" %in% cols) {
+      depmap_metadata <- depmap_metadata %>%
+        dplyr::mutate(stripped_cell_line_name = toupper(gsub("[^A-Za-z0-9]", "", .data$Cell_Line_Name)))
     } else {
       return(NULL)
     }
@@ -203,9 +223,8 @@ gimap_annotate <- function(.data = NULL,
   }
 
   ############################ Get TPM data ####################################
+  depmap_annotation_ok <- isTRUE(cell_line_annotate)
   if (cell_line_annotate) {
-    # This is used to flag things
-    ## get TPM and CN information (w/ option for user to upload their own info)
     metadata_file <- getOption("gimap_depmap_metadata_file")
     if (!is.null(metadata_file) && file.exists(metadata_file)) {
       depmap_metadata <- tryCatch(
@@ -235,71 +254,99 @@ gimap_annotate <- function(.data = NULL,
     }
 
     if (is.null(depmap_metadata)) {
-      message("Returning dataset without cell line annotation due to network error.")
-      return(gimap_dataset)
-    }
-
-    depmap_metadata <- normalize_depmap_metadata(depmap_metadata)
-    if (is.null(depmap_metadata) ||
+      message(
+        "DepMap metadata could not be loaded. ",
[…]
