### 2026-04-16 00fb82f
Address quick start changes
