+++ vignettes/quick-start.Rmd
@@ -16,7 +16,43 @@ knitr::opts_chunk$set(
 
 # Quick Start for gimap
 
-For more background on gimap and the calculations done here, [read here](https://github.com/FredHutch/gimap/blob/main/README.md)
+For more background on gimap and the calculations done here, [read here](https://github.com/FredHutch/gimap/blob/main/README.md).
+
+## About this example
+
+This Quick Start walks through a dual-targeting (paired-guide) CRISPR screen
+from start to finish using a small example dataset bundled through Figshare.
+
+**Study design of the example data.** The counts come from a paralog genetic
+interaction screen in HeLa cells using the pgPEN (paired guide RNAs for
+paralog genetic interaction mapping) library. Cells expressing Cas9 are
+transduced with a pooled library of paired-guide RNAs (pgRNAs) that knock out
+either one gene (single-target constructs, paired with a non-targeting
+control) or two paralogous genes simultaneously (double-target constructs).
+Cells are collected at a baseline timepoint (Day 0, representing the library
+composition before selection) and again after growth selection (Day 22),
+with three biological replicates at the later timepoint. Changes in pgRNA
+abundance between Day 0 and Day 22 reflect the fitness effect of each
+single- or double-knockout, and comparing observed to expected double-knockout
+fitness reveals genetic interactions (synthetic lethality or buffering).
+
+The example data are a subset of the screen described in Thompson *et al.*
+(2021), *Cell Reports* - [Discovery of synthetic lethal and tumor suppressor
+paralog pairs in the human genome](https://pubmed.ncbi.nlm.nih.gov/34469736/).
+Refer to that paper for the full experimental protocol, library composition,
+and discussion of the biological results.
+
+**How to map your own experiment onto this workflow.** You will need:
+
+- a counts table of pgRNA reads (rows = pgRNA constructs, columns = samples),
+- a table of pgRNA construct IDs (matching the row order of the counts), and
+- a sample metadata table (matching the column order of the counts) that
+  describes each sample - in this example, the timepoint (`day`) and
+  replicate (`rep`).
+
+The rest of the tutorial shows how these three pieces are assembled into a
+`gimap_dataset` object and run through QC, filtering, annotation,
+normalization, and genetic interaction scoring.
 
 ## Requirements
 
@@ -43,12 +79,15 @@ remotes::install_github("FredHutch/gimap")
 
 ## Loading needed packages
 
-```{r echo = FALSE, results = 'hide'}
-library(gimap)
-```
+Attach `gimap` and `dplyr` before running any of the code below. We use
+`suppressPackageStartupMessages()` to keep the rendered output tidy, but the
+calls themselves are essential.
 
-```{r echo = FALSE, results = 'hide'}
-library(dplyr)
+```{r}
+suppressPackageStartupMessages({
+  library(gimap)
+  library(dplyr)
+})
 ```
 
 ## Set Up
@@ -60,9 +99,30 @@ output_dir <- "output_timepoints"
 dir.create(output_dir, showWarnings = FALSE)
 ```
 
+Download the example counts table from Figshare. `get_example_data("count")`
+caches the file under `tempdir()` on first use and reads it in as a
+[tibble](https://tibble.tidyverse.org/):
+
 ```{r   eval = FALSE}
 example_data <- get_example_data("count")
 ```
+
+`example_data` is a tibble where each row is a paired-guide (pgRNA) construct
+and each column is either a construct-level identifier or a sample's read
+counts. The columns are:
+
+- `id` - unique identifier for the pgRNA construct (e.g. `CCNL1_CCNL2_pg1`).
+- `seq_1` - gRNA sequence targeting the first gene ("paralog A") in the pair.
[…]
