### 2026-04-16 FredHutch/gimap#114 @cansavvy
Addressed in https://github.com/FredHutch/gimap/commit/00fb82f8756c7a4148f7a9234cc78c7dfc66fb70
