### 2026-04-16 FredHutch/gimap#112 @cansavvy
Addressed this issue.

DepMap metadata (https://figshare.com/ndownloader/files/35020903) sometimes does not parse as a proper table (WAF/empty response, BOM on column names, or new column names like ModelID instead of DepMap_ID). And this problem was propagated to the rest of the steps.

gimap_normalize() failed with “No annotations are stored…” even though the real issue was DepMap, not missing gimap_annotate() in principle.

So now, normalize_depmap_metadata() rejects empty tables. We then strip UTF-8 BOM from the first column name, we map ModelID / model_id → DepMap_ID and we derive stripped_cell_line_name from Cell_Line_Name when needed.

These issues are often transient and work if you try again with figshare but now it will hopefully be more clear.
