### 2026-04-16 #9891 @cansavvy
> I have raised a few issues in the the [target repository](https://github.com/FredHutch/gimap).

Thanks for your reviews and patience. It's much appreciated. I am no longer in this position where I created this software originally and am no longer paid to maintain it so finding free time is tricky.

Here's a summary of the changes I did in response to @tomsing1 's very thoughtful review.

- License: Swapped to BSD-style text in LICENSE; DESCRIPTION → file LICENSE; updated docs license pages/links.
- README: CRAN vs GitHub install; DocToc refreshed.
- Example data: Session cache under tempdir(); removed risky options() paths; safe delete_example_data() (known files, allowed dirs); tests added; fixed roxygen so gimap_annotate docs attach to the right function.
- run_qc: Dropped unused plots_dir; updated template, vignettes, tests, docs.
- DepMap / annotate / normalize: More robust metadata parsing; graceful fallback when DepMap fails (still get annotations); optional CN; gimap_normalize auto turns off normalize_by_unexpressed if TPM flags missing; vignette note.
- Tests: Adjusted annotate test for new normalize behavior; run_qc skips without

All tests pass successfully locally.
