### 2026-05-01 #9891 @kalyanidhusia
@cansavvy

I began a code-based review by installing the package and running the quick-start workflow using modules `get_example_data()`, and also reviewed the normalization `gimap_normalize` and DepMap annotation `depmap_metadata` functions and quick-start vignette: https://fredhutch.github.io/gimap/articles/quick-start.html

1. How I tested (for context)

- I cloned the repo locally
- Installed using both: `standard install (CRAN / GitHub)` and `local install from cloned repo `
- Ran the quick-start workflow end-to-end
- Also ran a small comparison test around normalization behavior

    >> Let me know if it would be useful for me to share the small test scripts I used.

2. Small usability notes

- Input format clarity (setup_data)
I initially ran into input-shape errors when passing full example data directly. The workflow works well when following this pattern:
   ` select relevant columns > convert to matrix > derive pg_ids from same object `

Suggestion(i think It was pointed out before as well)

   - [ ] Add one short note in docs that `counts` should be a matrix
   - [ ] Maybe show `as.matrix()` explicitly in the example

code: https://github.com/FredHutch/gimap/blob/c8539aa0284c014c21c597ec237f7bb4eb377e0d/R/00-setup_data.R

3. Local install note (very minor)

- Installing from a cloned repo worked, but took a bit more troubleshooting than the standard install
- Suggestion:
     - [ ] Add 1–2 lines in README for local `install (e.g. remotes::install_local())`

4. Normalization behavior (small clarification request)

I ran a quick comparison:

default: `gimap_normalize(timepoints = "day")`
fallback: `normalize_by_unexpressed = FALSE `

For the example dataset, results were identical `(GI scores and FDR unchanged) `

suggestion

- [ ]     Add a short note explaining when this parameter is expected to matter, or
- [ ]     Mention if example dataset is intentionally insensitive

code: https://github.com/FredHutch/gimap/blob/c8539aa0284c014c21c597ec237f7bb4eb377e0d/R/04-normalize.R

5. Annotation / DepMap

The fallback behavior works smoothly. Just a small suggestion to clarify it for users:
   - [ ] Add a short explanation of fallback behavior in vignette

code: https://github.com/FredHutch/gimap/blob/c8539aa0284c014c21c597ec237f7bb4eb377e0d/R/03-annotate.R

Happy to help!
