+++ paper.bib
@@ -1,30 +1,16 @@
-
 @article{benjamini1995controlling,
-  title={Controlling the false discovery rate: a practical and powerful approach to multiple testing},
-  author={Benjamini, Yoav and Hochberg, Yosef},
-  journal={Journal of the Royal Statistical Society: Series B (Methodological)},
-  volume={57},
-  number={1},
-  pages={289--300},
-  year={1995},
-  publisher={Wiley Online Library}
+  title = {Controlling the false discovery rate: a practical and powerful approach to multiple testing},
+  author = {Benjamini, Yoav and Hochberg, Yosef},
+  journal = {Journal of the Royal Statistical Society: Series B (Methodological)},
+  volume = {57},
+  number = {1},
+  pages = {289--300},
+  year = {1995},
+  publisher = {Wiley Online Library},
+  doi = {10.1111/j.2517-6161.1995.tb02031.x},
+  url = {https://doi.org/10.1111/j.2517-6161.1995.tb02031.x}
 }
 
-@article{bioinformatics,
-    author = {Afiaz, Awan and Ivanov, Andrey A and Chamberlin, John and Hanauer, David and Savonen, Candace L and Goldman, Mary J and Morgan, Martin and Reich, Michael and Getka, Alexander and Holmes, Aaron and Pati, Sarthak and Knight, Dan and Boutros, Paul C and Bakas, Spyridon and Caporaso, J Gregory and Del Fiol, Guilherme and Hochheiser, Harry and Haas, Brian and Schloss, Patrick D and Eddy, James A and Albrecht, Jake and Fedorov, Andrey and Waldron, Levi and Hoffman, Ava M and Bradshaw, Richard L and Leek, Jeffrey T and Wright, Carrie},
-    title = {Best practices to evaluate the impact of biomedical research software—metric collection beyond citations},
-    journal = {Bioinformatics},
-    volume = {40},
-    number = {8},
-    pages = {btae469},
-    year = {2024},
-    month = {07},
-    abstract = {Software is vital for the advancement of biology and medicine. Impact evaluations of scientific software have primarily emphasized traditional citation metrics of associated papers, despite these metrics inadequately capturing the dynamic picture of impact and despite challenges with improper citation.To understand how software developers evaluate their tools, we conducted a survey of participants in the Informatics Technology for Cancer Research (ITCR) program funded by the National Cancer Institute (NCI). We found that although developers realize the value of more extensive metric collection, they find a lack of funding and time hindering. We also investigated software among this community for how often infrastructure that supports more nontraditional metrics were implemented and how this impacted rates of papers describing usage of the software. We found that infrastructure such as social media presence, more in-depth documentation, the presence of software health metrics, and clear information on how to contact developers seemed to be associated with increased mention rates. Analysing more diverse metrics can enable developers to better understand user engagement, justify continued funding, identify novel use cases, pinpoint improvement areas, and ultimately amplify their software’s impact. Challenges are associated, including distorted or misleading metrics, as well as ethical and security concerns. More attention to nuances involved in capturing impact across the spectrum of biomedical software is needed. For funders and developers, we outline guidance based on experience from our community. By considering how we evaluate software, we can empower developers to create tools that more effectively accelerate biological and medical research progress.More information about the analysis, as well as access to data and code is available at https://github.com/fhdsl/ITCR\_Metrics\_manuscript\_website.},
-    issn = {1367-4811},
-    doi = {10.1093/bioinformatics/btae469},
-    url = {https://doi.org/10.1093/bioinformatics/btae469},
-    eprint = {https://academic.oup.com/bioinformatics/article-pdf/40/8/btae469/58723814/btae469.pdf},
-}
 
 @misc{depmap2025,
   author = {{DepMap, Broad}},
