+++ paper.bib
@@ -18,6 +18,7 @@
   year = {2025},
   howpublished = {Dataset},
   url = {https://depmap.org},
+  doi = {10.1146/annurev-cancerbio-062722-024239},
   note = {Please note, you may need to update the release quarter depending on which version of the data you are using}
 }
 
