### 2026-06-29 1c88585
Align release metadata for JOSS archive
Co-authored-by: Cursor <cursoragent@cursor.com>
