+++ R/utils.R
@@ -20,7 +20,8 @@ utils::globalVariables(c(
   "mean_single_crispr_1", "expected_single_crispr", "double_crispr", "double_gi_score",
   "fdr", "lfc", "mean_expected_cs", "mean_gi_score", "mean_single_crispr",
   "expected_double_crispr", "p_val", "single_gi_score", "Rank", "broad_target_type",
-  "logfdr", "pointColor", "both", "mean_score", "gi_score", "Day05_RepA", "id"
+  "logfdr", "pointColor", "both", "mean_score", "gi_score", "Day05_RepA", "id",
+  "depmap_id", "ModelID", "model_id"
 ))
 
 #' Check if internet/URL is available
