+++ DESCRIPTION
@@ -4,24 +4,21 @@ Title: Calculate Genetic Interactions for Paired CRISPR Targets
 Version: 1.1.2
 Authors@R: c(
   person("Candace", "Savonen",
-         email = c("cansav09@gmail.com","csavonen@fredhutch.org"), role = c("aut", "cre")),
+         email = c("cansav09@gmail.com","csavonen@fredhutch.org"),
+         role = c("aut", "cre"),
+         comment = c(ORCID = "0000-0001-6331-7070")),
   person("Phoebe", "Parrish",
-         email = "pcrparrish@gmail.com", role = c("aut")),
+         email = "pcrparrish@gmail.com", role = c("aut"),
+         comment = c(ORCID = "0000-0002-3530-0105")),
   person("Kate", "Isaac",
-         email = "kisaac@fredhutch.org", role = c("aut")),
-  person("Howard", "Baek",
-         email = "howardbaek.fh@gmail.com", role = c("aut")),
+         email = "kisaac@fredhutch.org", role = c("aut"),
+         comment = c(ORCID = "0000-0001-8701-4197")),
   person("Daniel", "Grosso", role = c("aut")),
-  person("Siobhan", "O'Brien", role = c("aut")),
-  person("Alice", "Berger", role = c("aut")))
-Author: Candace Savonen [aut, cre],
-  Phoebe Parrish [aut],
-  Kate Isaac [aut],
-  Howard Baek [aut],
-  Daniel Grosso [aut],
-  Siobhan O'Brien [aut],
-  Alice Berger [aut]
-Maintainer: Candace Savonen <cansav09@gmail.com>
+  person("Marissa", "Fujimoto", role = c("aut")),
+  person("Siobhan", "O'Brien", role = c("aut"),
+         comment = c(ORCID = "0000-0003-4958-1080")),
+  person("Alice", "Berger", role = c("aut"),
+         comment = c(ORCID = "0000-0002-6538-2658")))
 Description: Helps find meaningful patterns in complex genetic experiments. First gimap takes data from paired CRISPR (Clustered regularly interspaced short palindromic repeats) screens that has been pre-processed to counts table of paired gRNA (guide Ribonucleic Acid) reads. The input data will have cell counts for how well cells grow (or don't grow) when different genes or pairs of genes are disabled. The output of the 'gimap' package is genetic interaction scores which are  the distance between the observed CRISPR score and the expected CRISPR score. The expected CRISPR scores are what we expect for the CRISPR values to be for two unrelated genes. The further away an observed CRISPR score is from its expected score the more we suspect genetic interaction. The work in this package is based off of original research from the Alice Berger lab at Fred Hutchinson Cancer Center (2021) <doi:10.1016/j.celrep.2021.109597>.
 License: file LICENSE
 URL: https://github.com/FredHutch/gimap
