### 2026-06-29 a4e7008
Tighten package build metadata for release
