### 2026-06-29 0659ed5
Prepare CRAN release 1.1.3
