+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Type: Package
 Package: gimap
 Title: Calculate Genetic Interactions for Paired CRISPR Targets
-Version: 1.1.2
+Version: 1.1.3
 Authors@R: c(
   person("Candace", "Savonen",
          email = c("cansav09@gmail.com","csavonen@fredhutch.org"),
@@ -20,7 +20,7 @@ Authors@R: c(
   person("Alice", "Berger", role = c("aut"),
          comment = c(ORCID = "0000-0002-6538-2658")))
 Description: Helps find meaningful patterns in complex genetic experiments. First gimap takes data from paired CRISPR (Clustered regularly interspaced short palindromic repeats) screens that has been pre-processed to counts table of paired gRNA (guide Ribonucleic Acid) reads. The input data will have cell counts for how well cells grow (or don't grow) when different genes or pairs of genes are disabled. The output of the 'gimap' package is genetic interaction scores which are  the distance between the observed CRISPR score and the expected CRISPR score. The expected CRISPR scores are what we expect for the CRISPR values to be for two unrelated genes. The further away an observed CRISPR score is from its expected score the more we suspect genetic interaction. The work in this package is based off of original research from the Alice Berger lab at Fred Hutchinson Cancer Center (2021) <doi:10.1016/j.celrep.2021.109597>.
-License: file LICENSE
+License: BSD_3_clause + file LICENSE
 URL: https://github.com/FredHutch/gimap
 BugReports: https://github.com/FredHutch/gimap/issues
 Depends:
