### 2026-06-29 9adc6e2
Update paper.md
