### 2026-07-01 a8b3cdf
Fix up names and funding based on reviews
