### 2026-07-05 #9891 @cansavvy
> Title, author list, version tag, and license match between the github repo and the zenodo archive. one thing to note is that the author names differ slightly between the two. @cansavvy  - you might change the author names on the zenodo archive to ensure that they match exactly those in the rendered pdf manuscript.

Ah thanks. I will fix this Monday when I'm at my computer.
